Use AWS SDK v2 if available for AWS credential fetching
To fetch AWS credentials from the environment:

  * Use AWS SDK v2 if available on the classpath
  * Otherwise, use the existing driver implementation

JAVA-4718
